Abstract profile picture section into reusable ProfilePicture component
